move tracing package from /pkg to /internal
